+++ src/redux/dialog/selectors.js
@@ -1,6 +1,8 @@
 // @flow
 import type { State } from "./types";
 
+export const open = (state: State) => state.open;
+
 export const type = (state: State) => state.type;
 
 export const artist = (state: State) => state.artist;