+++ src/redux/selectors.js
@@ -8,6 +8,8 @@ import type { State } from "./types";
 
 export const dialogArtist = (state: State) => dialog.artist(state.dialog);
 
+export const dialogOpen = (state: State) => dialog.open(state.dialog);
+
 export const dialogType = (state: State) => dialog.type(state.dialog);
 
 export const message = (state: State) => state.message;