+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix wrong button caption when artist creation dialog is closed.
+  ([#123](../../issues/123))
 
 ## [0.20.0] - 2019-07-19
 ### Added